# internal/api/openapi.yaml
openapi: "3.1.0"
info:
  title: 3GPP Server API
  description: |
    A JSON-to-NGAP/NAS translation server for conformance testing, fuzzing,
    and security assessment of 5G core networks.

    The server manages gNB and UE contexts, translates JSON requests into
    NGAP/NAS messages over SCTP, and returns decoded responses as JSON.

    ## Workflow

    1. **Create a gNB** — dials SCTP to the AMF, performs NG Setup.
    2. **Create a UE** — provisions credentials (K, OPc), computes SUCI.
    3. **Send NGAP messages** — builds NAS inside NGAP, sends over SCTP,
       returns the decoded AMF response.

    ## IE Control

    The JSON schema mirrors NGAP/NAS IEs 1:1. Every optional IE can be
    individually set, overridden, or omitted — enabling both correct
    procedures and deliberate deviations for fuzz testing.

    ## 3GPP References

    - NGAP: TS 38.413
    - NAS 5GS: TS 24.501
    - 5G Security: TS 33.501
  version: "0.2.0"
  license:
    name: BUSL-1.1
    identifier: BUSL-1.1
paths:
...
  /enb/{enb_id}/ue/{ue_id}:
    parameters:
      - $ref: "#/components/parameters/enb_id"
      - $ref: "#/components/parameters/ue_id"
    get:
      operationId: getENBUE
      summary: Get an eNB UE's state
      description: |
        Returns the UE's EPS state — its identities, S1AP ID pair, whether the NAS
        security context is active, the default bearer, and the established PDN
        connections.
      responses:
        "200":
          description: UE state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ENBUEStateResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: patchENBUE
      summary: Update an eNB UE's stored context
      description: |
        Overrides stored UE state before a send, so a step can be driven with
        altered credentials or identifiers. Only the fields present are changed.
        There is no slice (sst/sd) counterpart: slicing is 5GS-only (TS 23.501 §5.15).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                k: { type: string }
                opc: { type: string }
                amf: { type: string }
                sqn: { type: string }
                mme_ue_s1ap_id: { type: integer, description: "Force the stored MME UE S1AP ID." }
      responses:
        "204": { description: "UE updated." }
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteENBUE
      summary: Delete an eNB UE
      description: Removes the UE context from the eNB without signalling the MME.
      responses:
        "204": { description: "UE deleted." }
        "404":
          $ref: "#/components/responses/NotFound"
  /enb/{enb_id}/ue/{ue_id}/tunnel:
    parameters:
      - $ref: "#/components/parameters/enb_id"
      - $ref: "#/components/parameters/ue_id"
    get:
      operationId: getENBTunnel
      summary: Get the S1-U GTP-U tunnel state for a UE's EPS bearer
      description: |
        Returns the captured tunnel for the selected bearer (the default bearer
        when `ebi` is absent): the S-GW uplink endpoint (sgw_ip, ul_teid), the
        eNB downlink TEID (dl_teid), and the UE IP (ue_ip). Populated when the
        bearer was established on a GTP-U-enabled eNB.
      parameters:
        - name: ebi
          in: query
          schema: { type: integer }
          description: "EPS bearer identity. Defaults to the UE's default bearer."
      responses:
        "200": { description: "Tunnel state." }
        "400": { description: "No user-plane tunnel for the bearer." }
        "404":
          $ref: "#/components/responses/NotFound"
  /enb/{enb_id}/ue/{ue_id}/migrate:
    parameters:
      - $ref: "#/components/parameters/enb_id"
      - $ref: "#/components/parameters/ue_id"
    post:
      operationId: migrateENBUE
      summary: Relocate a UE context to a target eNB after S1 handover
      description: |
        Moves the UE context to the target eNB's association, modelling the UE
        arriving at the target after an S1 handover (TS 23.401 §5.5.1.2). The UE
        keeps its security context; its S1AP ID pair becomes the target's values.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target_enb_id]
              properties:
                target_enb_id:
                  type: string
                  description: "Store ID of the eNB to move the UE to."
                mme_ue_s1ap_id:
                  type: integer
                  description: "MME UE S1AP ID at the target (from the Handover Request)."
                enb_ue_s1ap_id:
                  type: integer
                  description: "eNB UE S1AP ID the target assigned to the UE."
      responses:
        "200":
          description: UE relocated.
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /enb/{enb_id}/s1ap:
    parameters:
      - $ref: "#/components/parameters/enb_id"
    post:
      operationId: sendENBS1AP
      summary: Send a non-UE-associated S1AP message (target-eNB side of S1 handover)
      description: |
        Sends a non-UE-associated S1AP message on the eNB's S1-MME association —
        the target-eNB side of S1 handover, where the eNB acts on a UE it does not
        yet own and addresses it by the S1AP ID pair the MME assigned.

        ## Supported message types

        - `handover_request_acknowledge` — Target-eNB HANDOVER REQUEST ACKNOWLEDGE
          (TS 36.413 §8.4.2), admitting the E-RABs in `admitted_erabs` with their
          downlink S1-U tunnels and rejecting those in `failed_erabs`. Send-only.
        - `handover_notify` — Target-eNB HANDOVER NOTIFY once the UE has arrived
          (TS 36.413 §8.4.3). Send-only.
        - `handover_failure` — Target-eNB HANDOVER FAILURE (TS 36.413 §8.4.2),
          making the MME abort the handover. Send-only.
        - `path_switch_request` — PATH SWITCH REQUEST (TS 36.413 §8.6.1) claiming a
          UE by its `mme_ue_s1ap_id` with a fresh `enb_ue_s1ap_id` and the `erabs`
          to switch. With `wait_for` the call blocks for the acknowledge, failure or
          Error Indication.

        ## Raw escape hatch

        `raw_s1ap_pdu` (hex) is the universal primitive: when set, `message_type`
        is ignored and the bytes are written verbatim onto the S1-MME association
        with no encoding or validation. With `wait_for` the call blocks for one of
        those downlink message types and returns it; otherwise it is
        fire-and-forget.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendENBS1APRequest"
      responses:
        "200":
          description: Sent (S1AP reply when the message expects one).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendENBNASResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        "502":
          $ref: "#/components/responses/BadGateway"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
...
  /gnb/{gnb_id}/ue/{ue_id}:
    parameters:
      - $ref: "#/components/parameters/gnb_id"
      - $ref: "#/components/parameters/ue_id"
    get:
      operationId: getUE
      summary: Get UE state
      description: Returns the full UE context including credentials, NGAP IDs, and session config.
      responses:
        "200":
          description: UE state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UEStateResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: patchUE
      summary: Patch UE state
      description: |
        Overrides any UE state field permanently. Useful for fuzzing — mutate
        credentials, NGAP IDs, or session config before sending the next message.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchUERequest"
      responses:
        "204":
          description: UE patched.
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteUE
      summary: Delete UE
      responses:
        "204":
          description: UE deleted.
        "404":
          $ref: "#/components/responses/NotFound"
...
  /gnb/{gnb_id}/ue/{ue_id}/migrate:
    parameters:
      - $ref: "#/components/parameters/gnb_id"
      - $ref: "#/components/parameters/ue_id"
    post:
      operationId: migrateUE
      summary: Move a UE's context to another gNB
      description: |
        Moves a UE's context (including its security context) from this gNB's
        association to another gNB's, modelling the UE arriving at the target
        after an N2 handover. The RAN/AMF UE NGAP IDs become the ones in use on
        the target, so subsequent UE-associated messages can be sent there.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target_gnb_id]
              properties:
                target_gnb_id:
                  type: string
                ran_ue_ngap_id:
                  type: integer
                  description: "RAN UE NGAP ID the UE uses on the target."
                amf_ue_ngap_id:
                  type: integer
                  description: "AMF UE NGAP ID in use on the target connection."
      responses:
        "200":
          description: UE moved.
        "404":
          description: gNB, UE, or target gNB not found.
...
components:
  parameters:
    gnb_id:
      name: gnb_id
      in: path
      required: true
      schema:
        type: string
    ue_id:
      name: ue_id
      in: path
      required: true
      schema:
        type: string
    enb_id:
      name: enb_id
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadGateway:
      description: SCTP or NGAP error communicating with the AMF.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Failed to build or encode the message.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    GatewayTimeout:
      description: Timed out waiting for AMF response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    S1APResponse:
      type: object
      properties:
        pdu_type:
          type: string
          enum: [initiating_message, successful_outcome, unsuccessful_outcome]
        message_type:
          type: string
          description: "Decoded message name (e.g. S1SetupResponse, S1SetupFailure)."
        raw_hex:
          type: string
          description: Full S1AP PDU as hex.
        s1_setup_response:
          type: object
          description: Present on an S1SetupResponse; served GUMMEIs and MME capacity.
        s1_setup_failure:
          type: object
          description: Present on an S1SetupFailure; cause and optional time-to-wait.
        unknown_ies:
          type: array
          description: ProtocolIEs the received message carries that its message type does not model, in wire order.
          items:
            type: object
            properties:
              id:
                type: integer
              criticality:
                type: string
                description: >-
                  reject, ignore, or notify: how a receiver must act on an IE it does not
                  comprehend (TS 36.413 §10.3.2). A value outside the ENUMERATED range is
                  rendered Criticality(n).
              value_hex:
                type: string
                description: Raw open-type value bytes as hex.
...
    ENBUEStateResponse:
      type: object
      properties:
        ue_id:
          type: string
        imsi:
          type: string
        ue_ip:
          type: string
          description: Assigned in the Attach Accept; empty before the attach completes.
        security_active:
          type: boolean
        mme_ue_s1ap_id:
          type: integer
          description: Set by the MME after the first S1AP exchange.
        enb_ue_s1ap_id:
          type: integer
        default_ebi:
          type: integer
          description: EPS bearer identity of the default bearer.
        bearers:
          type: array
          description: PDN connections beyond the default bearer (TS 24.301 §6.5.1).
          items:
            type: object
            properties:
              ebi:
                type: integer
              apn:
                type: string
              ue_ip:
                type: string
...
    SendENBS1APRequest:
      type: object
      properties:
        message_type:
          type: string
          enum: [handover_request_acknowledge, handover_notify, handover_failure, path_switch_request]
        raw_s1ap_pdu:
          type: string
          description: When set, hex S1AP PDU written verbatim onto the S1-MME association with no encoding or validation; message_type is ignored. Lets a test send any S1AP PDU, well-formed or not.
        wait_for:
          type: array
          items: { type: string }
          description: With raw_s1ap_pdu or path_switch_request, downlink S1AP message types to block for (empty = fire-and-forget).
        timeout_ms:
          type: integer
          description: With raw_s1ap_pdu + wait_for, how long to wait. Default 5000.
        mme_ue_s1ap_id:
          type: integer
          description: MME UE S1AP ID the MME assigned to the incoming UE (from the Handover Request).
        enb_ue_s1ap_id:
          type: integer
          description: eNB UE S1AP ID the target eNB assigns the UE; reused across handover_request_acknowledge and handover_notify. On path_switch_request, a fresh value the relocating eNB claims, staling the source association (TS 36.413 §10.6).
        erabs:
          type: array
          description: On path_switch_request, the E-RABs to switch to this eNB with their downlink S1-U endpoints.
          items:
            type: object
            required: [id]
            properties:
              id: { type: integer, description: "E-RAB ID." }
              dl_teid: { type: integer, description: "Target eNB downlink TEID (default synthesized)." }
              dl_ip: { type: string, description: "Target eNB S1-U address (default the eNB's own)." }
        admitted_erabs:
          type: array
          description: On handover_request_acknowledge, the E-RABs the target admits with their downlink S1-U endpoints.
          items:
            type: object
            required: [id]
            properties:
              id: { type: integer, description: "E-RAB ID." }
              dl_teid: { type: integer, description: "Target eNB downlink TEID (default synthesized)." }
              dl_ip: { type: string, description: "Target eNB S1-U address (default the eNB's own)." }
        failed_erabs:
          type: array
          items: { type: integer }
          description: On handover_request_acknowledge, E-RAB IDs the target rejects.
        cause:
          type: integer
          description: On handover_failure, override the radio-network Cause (default ho-failure-in-target).
        cell_id:
          type: integer
          description: On handover_notify, override the E-UTRAN cell identity of the UE's new location (default 1).
    SendENBNASResponse:
      type: object
      properties:
        s1ap:
          $ref: "#/components/schemas/S1APResponse"
        nas:
          type: object
          description: Decoded downlink EPS NAS message.
        mac_verified:
          type: boolean
          description: >
            Whether a protected downlink's NAS-MAC verified under the
            independently-derived keys (set for the Security Mode Command).
...
    UEStateResponse:
      type: object
      properties:
        id:
          type: string
        supi:
          type: string
        suci:
          type: string
        mcc:
          type: string
        mnc:
          type: string
        ran_ue_ngap_id:
          type: integer
        amf_ue_ngap_id:
          type: integer
          description: Set by the AMF after the first NGAP exchange.
        k:
          type: string
        opc:
          type: string
        amf:
          type: string
        sqn:
          type: string
        snn:
          type: string
          description: Serving Network Name (derived from MCC/MNC).
        dnn:
          type: string
        sst:
          type: integer
        sd:
          type: string
        protection_scheme:
          type: string
        routing_indicator:
          type: string
        imeisv:
          type: string
    PatchUERequest:
      type: object
      description: Any field present is overwritten in the UE context.
      properties:
        k:
          type: string
        opc:
          type: string
        amf:
          type: string
        sqn:
          type: string
        amf_ue_ngap_id:
          type: integer
        dnn:
          type: string
        sst:
          type: integer
        sd:
          type: string
        imeisv:
          type: string
...